Reorganized project, added a proper Makefile, and an install target
