+++ .gitignore
@@ -1,3 +1,6 @@
+# Ignore files generated by the build
+build_config.mk
+
 # Ignore build targets
 demo
 libhll.a
