+++ Makefile
@@ -1,25 +1,92 @@
-AR=ar
-CC=g++
-#CFLAGS=-Wall -Werror -g
-CFLAGS=-Wall -Werror -O3
-RANLIB=ranlib
+# Copyright 2015 The libcount Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License. See the AUTHORS file for names of
+# contributors.
 
-all: demo libhll.a
+# Uncomment exactly one of the lines labelled (A), (B), and (C) below
+# to switch between compilation modes.
 
-demo: demo.o libhll.a
-	$(CC) -o demo demo.o libhll.a -lssl -lcrypto
+# A: Production use (full optimizations)
+#OPT ?= -O3 -DNDEBUG
 
-libhll.a: bitwise_ops.o hll.o
-	$(AR) rcs libhll.a bitwise_ops.o hll.o
-	$(RANLIB) libhll.a
+# B: Debug mode, with full line-level debugging symbols
+OPT ?= -g2
 
-.cc.o:
-	$(CC) $(CFLAGS) -c $<
+# C: Profiling mode: optimizations, but w/debugging symbols
+#OPT ?= -O3 -g2 -DNDEBUG
+
+PREFIX ?= /usr/local
+
+# Warning Flags
+WARNINGFLAGS = -Wall -Werror
+
+# Detect what platform we're building on
+$(shell CXX="$(CXX)" TARGET_OS="$(TARGET_OS)" \
+	./build_config build_config.mk .)
+
+# Include the file generated by the previous line to set build flags, sources
+include build_config.mk
+
+AR = ar
+RANLIB = ranlib
+CXXFLAGS += -I. -I./include $(PLATFORM_CXXFLAGS) $(OPT) $(WARNINGFLAGS)
+COUNT_OBJECTS = $(COUNT_FILES:.cc=.o)
+TESTS = 
+DEV = demo
+
+# Targets
+all: libcount.a $(DEV)
+
+.PHONY:
+check: $(TESTS)
+	for t in $(TESTS); do echo "** Running $$t"; ./$$t || exit 1; done
 
 .PHONY:
 clean:
-	-rm -f demo *.o libhll.a *~ .*~
+	-rm -f */*.o build_config.mk *.a $(TESTS)
+
+demo: dev/demo.o libcount.a
+	$(CXX) $(CXXFLAGS) dev/demo.o libcount.a -o $@ -lcrypto
+
+.PHONY: install
+install: libcount.a
+	cp libcount.a "$(PREFIX)/lib"
+	mkdir -p "$(PREFIX)/include/count"
+	cp include/count/*.h "$(PREFIX)/include/count"
+
+libcount.a: $(COUNT_OBJECTS)
+	$(AR) rcs libcount.a $(COUNT_OBJECTS)
+	$(RANLIB) libcount.a
 
 .PHONY:
-count:
-	wc -l *.h *.cc
+linecount:
+	wc -l $(CPPLINT_SOURCES)
+
+.PHONY:
+lint:
+	$(LINT_TOOL) $(CPPLINT_SOURCES)
+
+.PHONY:
+neat: clean
+	-rm -f *~ .*~ */*~ ./include/*/*~
+
+.PHONY:
+reformat:
+	clang-format -i $(CPPLINT_SOURCES)
+
+# TODO(tdial): Add unit test targets
+
+# Suffix Rules
+.cc.o:
+	$(CXX) $(CXXFLAGS) -c $< -o $@
+
